+++ Generic Fantasy Game X/Assets/Scripts/OldCameraController.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 7d2f8c718081f4c4e8b7db280fcc011d
-timeCreated: 1526031966
+guid: a10352a2c43d80a49b6a718591631089
+timeCreated: 1526031912
 licenseType: Free
 MonoImporter:
   externalObjects: {}
