+++ Generic Fantasy Game X/Library/LastSceneManagerSetup.txt
@@ -1,4 +1,4 @@
 sceneSetups:
-- path: Assets/Scenes/CameraControls.unity
+- path: Assets/Scenes/HostileEntitiesNavigation.unity
   isLoaded: 1
   isActive: 1
